// frontend/lib/api/schema.d.ts
export interface components {
    schemas: {
// ...
        AssignmentStatusCounts: {
            draft: number;
            grading: number;
            review: number;
            completed: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}